requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0

# LangChain and related packages
langchain==0.3.25
langchain-openai==0.3.17
langchain-community==0.3.24
langchain-core==0.3.60
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.70
langsmith==0.3.42

# HTTP and API related (minimal set)
httpx==0.28.1

# AI and ML related (only what's used)
sentence-transformers==4.1.0
openai==1.81.0
transformers==4.52.2

# Search functionality
duckduckgo-search==8.0.2

# Browser automation
playwright==1.52.0
screeninfo==0.8.1

# Data processing and utilities
pydantic==2.11.4
pydantic-settings==2.9.1

# CLI dependencies (for browser_use/cli.py)
click
textual

# Local package - install with: pip install -e .
-e .